# openapi.yaml
openapi: 3.0.3
info:
  title: Кутёж — API основного пользовательского сценария
  version: 1.0.0
  description: >-
    Публичные проверки доступности и маршруты входа, профиля, мероприятий и ленты.
    Остальные доменные маршруты реализованы в backend/internal/httpapi/routes.go.
servers:
  - url: https://kutezh-social.ru
paths:
...
  /api/v1/events:
    get:
      summary: Получить мероприятия
      security: [{BearerAuth: []}, {SessionCookie: []}]
      parameters:
        - {name: city, in: query, schema: {type: string}}
        - {name: category, in: query, schema: {type: string}}
        - {name: starts_from, in: query, schema: {type: string, format: date-time}}
        - {name: starts_to, in: query, schema: {type: string, format: date-time}}
        - {name: price_max_rubles, in: query, schema: {type: integer}}
      responses:
        '200':
          description: Список мероприятий
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events: {type: array, items: {$ref: '#/components/schemas/Event'}}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
    post:
      summary: Создать мероприятие
      security: [{BearerAuth: []}, {SessionCookie: []}]
      parameters:
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateEvent'}
      responses:
        '201':
          description: Созданное мероприятие
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Event'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    BearerAuth: {type: http, scheme: bearer}
    SessionCookie: {type: apiKey, in: cookie, name: kutezh_session}
  responses:
    Error:
      description: Ошибка запроса
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: {type: string}
                  message: {type: string}
  schemas:
...
    Event:
      type: object
      required: [id, title, starts_at, location]
      properties:
        id: {type: integer}
        title: {type: string}
        starts_at: {type: string, format: date-time}
        location: {$ref: '#/components/schemas/Location'}
    CreateEvent:
      type: object
      required: [title, description, category, starts_at, location]
      properties:
        title: {type: string}
        description: {type: string}
        category: {type: string}
        starts_at: {type: string, format: date-time}
        location: {$ref: '#/components/schemas/Location'}
    Location:
      type: object
      required: [latitude, longitude]
      properties:
        venue_name: {type: string}
        address: {type: string}
        city: {type: string}
        latitude: {type: number, minimum: -90, maximum: 90}
        longitude: {type: number, minimum: -180, maximum: 180}
...